feat: category button 공용 컴포넌트 구현
