--- src/app/_home/_components/CategoryButton.tsx
@@ -1,30 +1,0 @@
-"use client";
-
-import { useState } from "react";
-
-export default function CategoryButton() {
-  const categories = ["전체", "오피스 스트레칭", "마인드풀니스"];
-
-  const [selectedCategory, setSelectedCategory] = useState<string>("전체");
-
-  // 카테고리 상태 업데이트
-  const handlecategoriesChange = (category: string) => {
-    setSelectedCategory(category);
-  };
-
-  return (
-    <div className="mb-3 mt-2 flex gap-3">
-      {categories.map((category, i) => {
-        return (
-          <button
-            key={category}
-            onClick={() => handlecategoriesChange(category)}
-            className={`h-[40px] w-auto rounded-lg border p-2 ${selectedCategory === category ? "bg-gray-900 text-white" : "bg-gray-200"}`}
-          >
-            {category}
-          </button>
-        );
-      })}
-    </div>
-  );
-}
