+++ src/components/SortButton.tsx
@@ -49,7 +49,7 @@ export default function SortButton({ cards, onSort }: SortButtonProps) {
       {/* 정렬 버튼 */}
       <button
         onClick={() => setSortDropdownOpen(!isSortDropdownOpen)}
-        className="mb-2 flex w-[110px] cursor-pointer rounded-xl border bg-gray-50 p-2 text-sm font-medium"
+        className="mb-2 flex w-auto cursor-pointer rounded-xl border bg-gray-50 p-2 text-sm font-medium"
       >
         <div className="mr-1">
           <SortIcon />
